Add support for posix-rename@openssh.com for the SFTP server side and fix tests accordingly.

+++ tests/stub_sftp.py
@@ -23,7 +23,7 @@ A stub SFTP server for loopback SFTP testing.
 import os
 import sys
 from paramiko import ServerInterface, SFTPServerInterface, SFTPServer, SFTPAttributes, \
-    SFTPHandle, SFTP_OK, AUTH_SUCCESSFUL, OPEN_SUCCEEDED
+    SFTPHandle, SFTP_OK, SFTP_FAILURE, AUTH_SUCCESSFUL, OPEN_SUCCEEDED
 from paramiko.common import o666
 
 
@@ -139,11 +139,23 @@ class StubSFTPServer (SFTPServerInterface):
     def rename(self, oldpath, newpath):
         oldpath = self._realpath(oldpath)
         newpath = self._realpath(newpath)
+        if os.path.exists(newpath):
+            return SFTP_FAILURE
         try:
             os.rename(oldpath, newpath)
         except OSError as e:
             return SFTPServer.convert_errno(e.errno)
         return SFTP_OK
+
+    def posix_rename(self, oldpath, newpath):
+        oldpath = self._realpath(oldpath)
+        newpath = self._realpath(newpath)
+        try:
+            os.rename(oldpath, newpath)
+        except OSError as e:
+            return SFTPServer.convert_errno(e.errno)
+        return SFTP_OK
+
 
     def mkdir(self, path, attr):
         path = self._realpath(path)
